Change elementwise output mapping to allow use of mapping of a constant tensor
Summary:
It seems reasonable that a constant could be used to determine the output mapping for an elementwise op.
Suppose a scalar is added to a large, specifically mapped constant - the output could be helpfully mapped the same as that constant, rather than fall to the default of being mapped linearly starting from zero.

This was not allowed by the code before this diff - the constant was rejected as a candidate for determining the mapping

Ref T62710

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Subscribers: godfrey.da.costa

Maniphest Tasks: T62710

Differential Revision: https://phabricator.sourcevertex.net/D67401