Deduplicate code that selects the ConvPartial1x1 vertex
Summary:
There were 3 slightly different implementations of the code used to select the ConvPartial1x1 vertex instead of the nx1 vertex.
The most important differences were in how the planner tried to detect if multiple worklist entries would be used; it used the input field shape rather than the output field shape. In preparation for tile- and vertex-level expand dims, which changes the input field shape, this patch unifies all 3 implementations and uses the output field shape to detect multiple worklist entries per worker.

Test Plan:
Existing tests to check nothing broke.
Added an assert that the number of partitions per worker is 1 when using the 1x1 vertex.
I added a print to low-level vertex selection that would print to stdout whenever the 1x1 vertex was selected and found all the 1x1 tests printed it the same number of times with and without this patch.

Reviewers: tomm, samk, godfrey.da.costa, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T38785

Differential Revision: https://phabricator.sourcevertex.net/D67589