Allow serial splits of the output channels of a convolution by non-factors
Summary:
Allows a non-factor number of serial splits of the output channels of a convolution using an approach outlined in the T22373 comments.

Also reworks `createWeightsImpl` and `createInputImpl` to use the input/output channel group sizes defined in the plan to create an initial tensor. This is then trimmed down to the right size, since the actual channel counts may not be multiples of the group sizes. This seems to make it easier for `mapWeights` to map serial slices consistently so they play nicely with `dynamicSlice`.

If the split count isn't a factor of the channel count then the split size is calculated by effectively rounding it up to the next multiple of the split count and then dividing. So, if the split count is `n`, we then have `n-1` equally sized splits, and then one smaller split.

The first `n-1` are processed as usual in loop where:
 - `dynamicSlice` copies the corresponding weights to a slice tensor.
 - The weights slice is convolved with the input to produce an output slice.
 - `dynamicUpdate` copies the output slice to the output tensor.

{F2081958}

The last split is then processed after the loop:
 - The weights are copied from the weights tensor to the same slice tensor used by the dynamic slice above.
   - Note that the weights won't fill the slice tensor, so there'll be output channels left over from the loop.
 - The weights slice is convolved with the input using the same code as the rest of the slices.
   - This allows us to save some memory, but means cycles are spent processing the left over output channels described above.
 - Only the channels that we need are copied out of the output slice to the output tensor.

{F2081960}

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk, godfrey.da.costa, tomm, richard

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Subscribers: timr

Maniphest Tasks: T22373

Differential Revision: https://phabricator.sourcevertex.net/D65408