Add optional support for vertex level expand dims
Summary:
The expand dims transform creates overlapping slices of input data which when connected to the ConvPartial vertex as a single tensor Poplar will broadcast the data to a new tensor to remove the aliasing. However, this broadcast can be avoided if the slices of input data are connected as multiple tensors to the vertex, saving cycles and temporary memory.

The expand dims transform can be performed at the system or tile level, however the tile level was not considered in the planner and only useable via a constraints file. This patch adds a conv option, `enableTileLevelExpandDims`, that when set to true lets the planner try tile-level expand dims candidates and adds a "vertex-level" expand dims transform that connects the slices of inputs as multiple tensors to the vertex. The tile-level transform will defer to the vertex-level whenever possible. The tile-level and system/ipu-level expand dims are considered separately by the planner to limit the planning workload.

Note that this patch requires follow-on work to make the conv modelling accurate and support convolutions with fewer input channels (hence the optimisation is currently disabled by default via the conv option).

Test Plan:
Existing tests pass.
Added unit tests pass. The added tests check for no input rearrangement in known-good cases.
Profiled some tests and observed both cycle and memory improvements when the transform is enabled.

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm, samk, godfrey.da.costa

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, godfrey.da.costa

Maniphest Tasks: T38785

Differential Revision: https://phabricator.sourcevertex.net/D66020