Improved ConvPartial1x1Out quarter,half vertices
Summary:
Changed to not overread after some testing with vertex verifier.  Rearranged to improve the speed of the fast path which leads to the main loop body, and avoid always adding to the loop count.  Updated cycle estimates to match

Ref T49867

Test Plan: Local execution of quarter tests

Reviewers: oleksiik, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: oleksiik, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T49867

Differential Revision: https://phabricator.sourcevertex.net/D66984