Automate measuring the accuracy of the planner
Summary:
The workflow for testing the accuracy of the planner is:

- Pick a workload for the convolution.
- Run it with profiling and dump the plan to debug logging
- Open with profiler and manually inspect the profile
- Compare the plan (from debug logging) to the profile visually.

This patch attempts to automate this process with the `test-planner`
options added to the `single_conv_layer`. The new functionality skips the manual
profiling steps by reading the profile data directly using the pva library.

Test Plan:
Ran many commands with debug logging
and verified that the estimated cycles matched what was dumped to the log
and verified that the actual cycles matched the Convolution steps in the
Graph Analyser. Note that there were some discrepancies with per-tile reporting
however the total cycles always matched, and the total memory usage looked
correct as well when compared to the not-always-live variables.

Reviewers: samk, tomm, godfrey.da.costa, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T38785

Differential Revision: https://phabricator.sourcevertex.net/D69280