Allow vertex expand dims to be used on non-contiguous tensors
Summary:
It can still be beneficial to defer to the vertex level expand dims
even when there will be some rearrangement due to acts not being
contiguous because it can still be less rearrangement than would
otherwise happen.

Test Plan:
Observed the vertex expansion being used in a constrained wu pass,
and found it was beneficial to temporary memory despite input padding
being applied (and thus causing the tensor to not be contiguous).

Reviewers: tomm, godfrey.da.costa, samk, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T38785

Differential Revision: https://phabricator.sourcevertex.net/D68958