Allow a non-factor number of serial splits of the input channels of a convolution
Summary:
Allows the number of serial splits of the input channel axis to be a non-factor of the total number of channels

The split size is effectively computed by rounding up the number of channels to the next multiple of the number of splits, and then dividing. So, if we have `n` splits, then we have `n-1` equally sized splits and one smaller one.

The first `n-1` splits are handled as usual by a loop:
 - Use dynamic slice to copy the current input and weight splits to slice tensors
 - Perform the convolution of the input and weight tensor slices to compute an output slice
 - Add the output slice to the output tensor (or just copy them over if this the first iteration of the loop)

The last split is then handled as a special case:
 - Copy the inputs and weights to the slice tensors described above
 - Zero the remaining input channels of the input or weights slice (whichever is smaller)
 - Add the output slice to the output tensor

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm, godfrey.da.costa

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Maniphest Tasks: T22373

Differential Revision: https://phabricator.sourcevertex.net/D68112