Add an 'identicalLayout' function to poplibs_test utils
Summary:
Ref T61567.

This will be useful for testing for T61567.

This checks that two tensors have an identical layout, meaning using either of the 2 tensors in an operation will produce the same result once lowered by Poplar.
The actual check is checking for equivalency of tile mapping, ordering of elements mapped to each tile, and aliasing.

Reviewers: janekvo, samk, richard, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: richard, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T61567

Differential Revision: https://phabricator.sourcevertex.net/D69613