[popops] Refactor implementation of multiSlice
Summary:
Split internal functions into parts which can be reused.

Motivation:
There is a method to create an input Tensor for a multiSlice (createSliceTensor) but
no method for using this tensor when `numIndices > 1`. For `numSlices = 1` there
is the method `dynamicSliceWithOutput`. I think we should have a method
`multiSliceWithOutput` for `numIndices > 1`. The factorization in this diff makes
implementing such a method easier.

Test Plan: No new tests added, as this is just an internal refactor.

Reviewers: tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, godfrey.da.costa

Reviewed By: tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, godfrey.da.costa

Maniphest Tasks: T65723

Differential Revision: https://phabricator.sourcevertex.net/D70231