Remove unnecessary use of output field
Summary:
This logic was likely originally trying to avoid calling
getMaxInputRangeSize as an optimisation in the case where
expand dims or flattening is present (as for an expanded dimension the
input field size of the dimension is the same as the output
field size of the corresponding dimension, and similarly for a flattened
dimension because to be flattened the kernel size will be 1).

However we support some transforms for which this isn't true,
so it makes the code rather surprising. Deleting for simplicity.

Test Plan:
Existing tests.
No changes to some expand dims plans.

Reviewers: tomm, godfrey.da.costa, samk, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T38785

Differential Revision: https://phabricator.sourcevertex.net/D71176