Using vertex vertifier to test ConvPartial1x1Out quarter,half vertices
Summary:
Added 4 test files for ConvPartial vertices with input quarter, output half, using `testConfig` in vertex verifier
Enabled cpp codelets for Ipu21 so that they can be verified.
The purpose of the tests is to check for overread when striding in, out in various directions.  This revealed bugs which are fixed.

Ref T49867

Test Plan: The tests that are added...

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oleksiik

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oleksiik

Maniphest Tasks: T49867

Differential Revision: https://phabricator.sourcevertex.net/D69367