Cast float to/from quarter directly, add castWithOutput (So that we can support a withoutput option)
Summary:
Casting float  to quarter and quarter to float.  Uses an intermediate half type and so 2 codelets in sequence.  This cannot be implemented when using a computeset API, (which should probably be deprecated anyway), and that's the only one that provided a with output capability.  So added castWithOutput

Fixes T68090

Test Plan: Specific tests added for the new function

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oleksiik, ranjayc

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, ranjayc

Maniphest Tasks: T68090

Differential Revision: https://phabricator.sourcevertex.net/D73903