Disable vertex-level expand dims due to large planning time increase.
Summary: This reverts commit 3b32a04ee0ef.

This caused planning time to jump from 30->50 seconds in a TF1 CNN model (T68715).
Also from a quick scan of the benchmarks it didn't seem to cause any significant
(non-wobble) improvements in throughput or latency. Could be that it's not being
used very much due to being disabled when input padding is present. Anyway, revert
until we've optimised the planner a bit and fixed the input padding case.

Test Plan: revert-hammer

Reviewers:

Subscribers: