Add a static sparse matrix multiplication that supports block size of 1, 4, 8 and 16
Summary:
Add a static sparse matrix multiply for fixed set of block
sizes. Extension to any block size doesn't require a
change to the API or the partitioner as it would just
mean a change in the block size. The API supports
the CSR/CSC/COO representations and uses a
custom partitioner to split work across threads
and tiles with a pre-processing state that permute
rows/columns.

Ref T&0839

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm, timp

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Subscribers: yanz, douglaso, valeriuo

Maniphest Tasks: T67386

Differential Revision: https://phabricator.sourcevertex.net/D70646