Casts with single contiguous region per tile should select a single worker rather than MultiVertex
Summary:
Cherry pick 4c6559bc2d006df from master

The ratio of the overhead of splitting work for a multi-vertex to
the cost of doing the actual cast itself is significant when small
amount of work is done. It would be cheaper to use a single worker
in such cases. Use a threshold based on the source and destination
type such that a single worker vertex is selected rather than a
multivertex.
A simplistic model to compare cycles against could be used. Ideally,
cost modelling should be used to make a better decision, but it may
be unnecessary overhead to do so but could be done in the future when
modelling is better.

!ci_branch_mk2

Test Plan: Verified that correct vertex is selected for small sizes

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, timp

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, timp

JIRA Issues: POPLAR-342

Differential Revision: https://phabricator.sourcevertex.net/D81056