Use vector width based on the actual types used in cast to decide grain size
Summary:
Cherry-pick 2cc6b271831ffe37d9a8ab2caf8979fbfae5483b

The grain size is incorrectly chosen to be the number of elements
in FLOAT rather than based on the actual types used in the cast.
That doesn't have a significant impact for types other than 8-bit
types like QUARTER, CHAR.

Set the grain size to be a maximum between the input and output
types in the cast as it will avoid rearranging copies owing to
the alignment constraints on the Cast vertex edges.

!ci_branch_mk2

Test Plan: Covered by existing tests

Reviewers: saml, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: saml, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

JIRA Issues: POPLAR-367

Differential Revision: https://phabricator.sourcevertex.net/D81388